fix: entropy topsis
